Updated `Rating` model to use `@Positive` validation for `orderNumber`, changed related templates for proper input types, updated `schema.sql` to use `INT` for `orderNumber`, and added `@Positive` validation for `bidQuantity` in `BidList`.

+++ src/main/resources/DB_scripts/schema.sql
@@ -74,7 +74,7 @@ CREATE TABLE rating (
   moodysRating VARCHAR(125),
   sandPRating VARCHAR(125),
   fitchRating VARCHAR(125),
-  orderNumber tinyint,
+  orderNumber INT,
 
   PRIMARY KEY (rating_id)
 );
